Updated DeleteTeller

+++ cap2/src/main/java/com/cecil/teller/DeleteTeller.java
@@ -1,23 +1,38 @@
 package com.cecil.teller;
 
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 
 import com.cecil.connection.Connections;
 
 public class DeleteTeller {
-        public static void delete(String tname, String tpass) {
+    public static void delete(String tname, String tpass) {
         try {
-            String sql = "delete from teller where tname = ? and tpass = ?";
+            String sql = "select * from teller where tname = ? and tpass = ?";
             PreparedStatement pstmt = Connections.openConn().prepareStatement(sql);
 
             pstmt.setString(1, tname);
             pstmt.setString(2, tpass);
 
-            pstmt.execute();
+            ResultSet r = pstmt.executeQuery();
 
-            System.out.println("=========================TELLER " + tname
-                    + " DELETED Successfully!===============================");
+            if (r.next()) {
+                String sql1 = "delete from teller where tname = ? and tpass = ?";
+                PreparedStatement pstmt1 = Connections.openConn().prepareStatement(sql1);
+
+                pstmt1.setString(1, tname);
+                pstmt1.setString(2, tpass);
+
+                pstmt1.execute();
+
+                System.out.println("=========================TELLER " + tname
+                        + " DELETED Successfully!===============================");
+
+            } else {
+                System.out.println("===================TELLER " + tname + " was not found."
+                        + " No Teller was DELETED !=========================");
+            }
         } catch (SQLException se) {
             System.out.println(se.getMessage());
         } finally {
